change fix for 3.2
even slashEquivocation() burns 50% of the MIN_PENALTY, an adversary can repeatedly call the function to self-slash and recover 50% of their collateral. If they are subject to slashing from a protocol with > 50% of the collateral as penalty, they can use slashEquivocation() to reduce the penalty to 50%. To mitigate we can only allow an operator to be slashed from equivocation once. This also allows us to remove the slashedSlots mapping and remove the permutation checks within slashEquivocation()