Fix Issues From Audit (#53)
* Registry issue fixed

* minor

* minor

* minor

* Update Registry.sol

* Update Registry.sol

* Update Registry.sol

* ran forge fmt

* merge into optimizations branch

* forge fmt

* remove dust check

operator.collateralWei stores wei so we don't have this issue anymore

* hash owner address with tree root instead of into each leaf

* remove inline comments for fix to 3.3

* remove searcher compensation

Acknowledge 3.7, protocols are meant to cover these costs as self-slashing is a prevalent issue that undermines searcher compensation

* add checks for operator deleted

* change fix for 3.2

even slashEquivocation() burns 50% of the MIN_PENALTY, an adversary can repeatedly call the function to self-slash and recover 50% of their collateral. If they are subject to slashing from a protocol with > 50% of the collateral as penalty, they can use slashEquivocation() to reduce the penalty to 50%. To mitigate we can only allow an operator to be slashed from equivocation once. This also allows us to remove the slashedSlots mapping and remove the permutation checks within slashEquivocation()

* fix 3.4: add _rewardAndBurn()

Since burning 100% - MIN_COLLATERAL is extreme for what could be a mistake, it has been changed to 50% burned, 50% to challenger like the slashEquivocation() case. Then the operator has to call claimSlashedCollateral() to reclaim their remaining collateral. Since both functions share the same logic, a _rewardAndBurn() function was added.

* move owner back into leaf when merkelizing and fix tests

* add owner to leaf hashing in _verifyDelegation()

* fix SlashRegistrationTester tests

* fix slashing tests

---------

Co-authored-by: Rachit2323 <atreyrachit23@gmail.com>
Co-authored-by: Rachit2323 <76684704+Rachit2323@users.noreply.github.com>
Co-authored-by: Irfan Shaik <YOUR_EMAIL>
Co-authored-by: JasonVranek <jasonvranek@gmail.com>